HW2/requirements.txt
lightly==1.2.13
matplotlib==3.3.4
numpy==1.19.5
Pillow==9.1.0
pytorch_lightning==1.6.1
scikit_learn==1.0.2
torch==1.8.1
torchvision==0.9.1
